Degenerified REST resources [ref #12]
